Separate API from worker
