+++ docs/conf.py
@@ -159,7 +159,7 @@ html_theme_options = {}
 # The name for this set of Sphinx documents.  If None, it defaults to
 # "<project> v<release> documentation".
 try:
-    from environment_provider import __version__ as version
+    from environment_provider_api import __version__ as version
 except ImportError:
     pass
 else:
