+++ src/entry.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-exec gunicorn environment_provider.webserver:FALCON_APP \
+exec gunicorn environment_provider_api.webserver:FALCON_APP \
 	--name environment_provider \
 	--worker-class=gevent \
 	--bind 0.0.0.0:8080 \
